refactor: remove unused product and user model extensions, update TypeScript configuration
