+++ packages/database/tsconfig.json
@@ -7,7 +7,7 @@
     "composite": false,
     "outDir": "dist"
   },
-  "include": ["src/**/*.ts", "models/**/*.ts", "prisma.config.ts"],
+  "include": ["src/**/*.ts", "prisma.config.ts"],
   "exclude": [
     "node_modules",
     "dist",
